Don't subselect data before passing to `extend`
This results in between 2-4x performance improvement, depending on
aggregation used, due to the reductions in pandas overhead (we were
suffering due to allocations and poor cache performance). This also
slightly improves resilience, as we're explicitly doing bounds checks in
the numba kernel, instead of trusting that everything was inbounds.
Wins all around.