First pass at redesign of aggregation outputs
Previously raw dynd arrays were output. This was problematic, as some of
the metadata that is known when the aggregate was created had to be
dropped, as there was no way to attach it to the dynd array. Non-array
represented data (such as categorical counts) had to be "hacked" on as
well, which wasn't as nice as it could be.

To solve these problems, we now return objects of type `Aggregate`,
which can contain arbitrary data, as well as the `x_axis` and `y_axis`
attributes (can be thought of as indexes along each dimension).

As an added benefit, these can now contain whatever methods we deem fit
to attach to aggregates, which can lead to a nicer api overall. This
will be accomplished in a later PR.