Remove fastparquet patch
This is probably happening due to API differences. In fastparquet, the has_nulls kwarg attribute can be a boolean, a list of strings, or None. The dask version accepts an integer. So for now I will not worry about this detail, and perhaps revisit or file an issue after the perf work is done