Remove usecols option
The keyword-argument "usecols=[..]" was causing both Pandas and Dask to ignore the index column of CSV files while reading them in, effectively shifting all columns one position to the left. In the case of writing out to parq, fastparquet (for Pandas) was flexible-enough to recover, however the Dask wrapper around it was not, which caused an error. I will try to make a minimal testcase tomorrow morning and file an issue with the Dask project.