Fix minor issue
Fix an issue which pops up whenever timed_read() gets called outside of the main() function. This happens because we use a global Parameters singleton, which is assumed to have been initialized by the time timed_read() is executed, but this is not the case when we call timed_write() from the command-line, because timed_write() also calls timed_read().