Upgrade packages to improve perf and add categoricals feature
Packages that were upgraded: numpy 1.11->1.12 (and all packages that depend on it, including pandas), fastparquet+dask -> master (for automatic detection of categoricals in dd.parquet), and python-snappy -> master (so that the GIL can be released). These upgrades were to ensure that the performance numbers reflect the latest improvements, and that categoricals are automatically detected when reading parquet files. Note that the "channels" section in filetimes.yml has also been removed, to preserve backwards-compatibility with defaults-channel packages.