Add dd-force-load option
Add the command-line option to force-load the Dask dataframe on read. This allows us to cycle through the caching/loading combinations to see whether that is causing a perf issue. Also, add some DEBUG messages to the print statements affected by the --debug flag, to make the debug output easier to read