Replace census.zip file in S3, add (some) memory profiling
census.zip needed updating in S3, as it contained an h5 file instead of a parq file. It now contains a parq file (uncompressed) that was written out with a categorical dtype for the "race" column. Also, filetimes.py now has some rudimentary memory profiling features which measure a) the memory of the overall process (max resident set size), and b) the memory of the dataframe after it is read. Finally, a CSV file detection issue is fixed (it was happening when dask wrote out the CSV files).