Fix issues with csv benchmarks
The "filetimes" benchmarks have had some issues with the CSV files, due to the fact that dask writes out the csvs as separate files (after partitioning), but pandas writes the csv data as a single larger file. The file sizes should now be reported properly, and the benchmarks will be based on loading all CSVs in either configuration (concatenating the resulting dataframes together with pandas).