Fix exclusive upper bound for line aggregation
The previous adjustment for the exlusive upper bound failed to account
properly for bounds with float values. This caused more pixels than
necessary to be clipped.

Instead of performing the upper bound check when we clipped a line
segment to the bounding box, we now check at a lower level when we
translate a line segment to the actual pixels. This gives us greater
control without dealing with float coordinates; at that level we are
working with integer coordinates.

Fix #342