Pin xarray version. Make py3.4 support optional. Update cvs.raster API
Pin xarray version to 0.9.6 so that the rasterio backend can be relied upon. Make Python 3.4 support optional to the Travis-CI build, since conda-forge no longer supports it (but might again in the future). Skip the tests that rely on the use_overviews kwarg in cvs.raster since it is no longer implemented. Update the cvs.raster() API to include upsample_method and downsample_method, to allow for finer-grained control during resampling.