Reduce memory profile of weighted edges
Previously I add a weight column with a constant value if the required
internal dataframe did not have a weight column. For directly connected
edges and bundled edges without a weight column, this caused the memory
profile to increase.

I refactored the code to reduce this profile and pass the necessary
parameters to the other helper methods.