Denormalize edge coordinates after bundling
Previously, we returned edges that used normalized coordinates. When
layering the non-normalized nodes onto the normalized edges, the result
was misaligned.

We now denormalize the edges coordinates only in hammer_bundle. The
directly_connect_edges method is simplified by not normalizing the
nodes in the first place.

Thus, performance will decrease for hammer_bundle, but will improve for
directly_connect_edges.