Fix upper-bound bin error for auto-ranged data (#459)
Previously, points falling on the upper bound of the plotted area were excluded from the plot, which was consistent with the behavior for individual grid cells, but which was confusing and misleading for the outer boundaries.  Points falling on the very outermost boundaries are now folded into the final grid cell, which should be the least surprising behavior. Fixes #457.