Filled area support (#734)
* gitignore updates

* Add trapezoid rendering functions and tests

Adds function to render trapezoids with parallel edges that are aligned with the y-axis. These are the primitives that will be used to create area glyphs.

* Remove trailing white space

* Add canvas.area function for rendering area plots

This function relies on the new AreaToZero and AreaToLine glyphs.  AreaToZero handles area plots between a line and y=0.  AreaToLine handles area plots between two lines that share the same x-axis coordinates.

* Add area-to-zero and area-to-line examples to the Timeseries user guide

* prevent y-index out of bounds when stack=False

* Add axis to class/function names

* WIP implementation of axis argument for area glyph

* Fix area to line pixel mapping

* Parametrize unit tests for all area specification formulations

* Canvas.area docstring and examples

* Update timeseries area example